update vLLM test
