+++ tests/core/language_model/vllm/test_vllm_custom_template.py
